pyproject.toml
[project]
name = "ynstock"
version = "0.1.0"
description = "股票行情数据服务"
requires-python = ">=3.11"
dependencies = [
    "akshare>=1.18.30",
    "baostock>=0.8.9",
    "fastapi>=0.133.1",
    "pandas>=3.0.1",
    "uvicorn>=0.41.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "jqdatasdk>=1.9.8",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "ipykernel>=7.2.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/username/ynstock"
Documentation = "https://github.com/username/ynstock#readme"
Repository = "https://github.com/username/ynstock.git"
